requirement.txt.txt
asgiref==3.7.2
Django==4.1.12
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
pymongo==4.5.0
pytz==2023.3.post1
sqlparse==0.2.4
tzdata==2023.3